+++ packages/edit-site/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Enhancements
 
--   Fluid typography: add configurable fluid typography settings to theme.json ([#42489](https://github.com/WordPress/gutenberg/pull/42489)).
+-   Fluid typography: add configurable fluid typography settings for minimum font size to theme.json ([#42489](https://github.com/WordPress/gutenberg/pull/42489)).
 
 ## 4.19.0 (2022-11-16)